pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = "src/factories.py"
pretty = true
show_error_context = true
show_error_codes = true
strict = true
warn_unused_ignores = true

[tool.poetry]
authors = ["Bryant Finney <bryant.finney@outlook.com>"]
description = "An example of a real-world DS&A problem"
name = "python-dsa"
packages = [{ include = "*.py", from = "src" }]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.8"
ipykernel = "^6.29.3"
typing-extensions = { version = "^4.10.0" }
sqlalchemy = "^2.0.28"
jupyterlab = "^4.1.5"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"
petname = "^2.6"
mypy = "^1.9.0"
ruff = "^0.3.2"

[tool.poetry.scripts]
create-test-db = "factories:create_test_db"
create-prod-db = "factories:create_prod_db"

[tool.pyright]
ignore = ["src/factories.py"]

[tool.ruff]
fix = true
line-length = 100
show-fixes = true
show-source = true

[tool.ruff.format]
quote-style = "single"

[tool.ruff.isort]
known-first-party = ["db", "factories", "models"]

[tool.ruff.lint]
select = [
  "B",
  "C90",
  "E1",
  "E4",
  "E7",
  "E9",
  "ERA",
  "F",
  "G",
  "I",
  "N",
  "PL",
  "PT",
  "RET",
  "RUF",
  "TID",
  "UP",
  "W"
]

[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"

[tool.ruff.mccabe]
max-complexity = 5

[tool.ruff.lint.pydocstyle]
convention = "google"
